Revert "Merge pull request #2933 from FloPinguin/fix-nation-loading"
This reverts commit dba04027df723880c096f0d3b975e831425f63fd.
