+++ src/client/styles.css
@@ -546,6 +546,7 @@ label.option-card:hover {
   flex-wrap: wrap;
   gap: 8px;
   justify-content: center;
+  padding: 0 16px;
 }
 
 /* News Button Notification */
